Switch console.log calls
The biggest remaining problem is that it can't seem to write the user
relationship to the notebook.nedb file. For the master branch, a notebook
record will have a k/v pair like "user":"1" to indicate belonging to the
user whose id is 1. The JsonApi branch does not write this properly in
app/notebook/route.js' createRecord call on line 10, despite recognizing
the user via

  this.controllerFor('application').get('user')

in a console.log call on line 9. Still investigating.